Refactor old threads helper file
